refactor to separate components and utils
