+++ index.html
@@ -5,7 +5,7 @@
     <meta charset="UTF-8" />
     <meta name="viewport"
           content="width=device-width, initial-scale=1.0" />
-    <title>FINOS Calendar</title>
+    <title>FINOS Event Calendar</title>
   </head>
 
   <body>
