+++ src/components/Header.jsx
@@ -1,0 +1,9 @@
+const Header = () => {
+  return (
+    <div className="content">
+      <h1 className="logo">FINOS Event Calendar</h1>
+    </div>
+  );
+};
+
+export default Header;
